refactor: flat project folder structure
