--- src/objects/station/station_snapshot.rs
@@ -1,11 +1,0 @@
-use rocket::time:: OffsetDateTime;
-
-use crate::objects::{subscriber::Subscriber, track::track::Track};
-
-pub struct StationSnapshot {
-    pub name: String,
-    pub current_track: Track,
-    pub subscribers: Vec<Subscriber>,
-    pub created_on: OffsetDateTime,
-    pub duration_secs: f64, 
-}
