--- src/objects/subscriber.rs
@@ -1,5 +1,0 @@
-// estrutura para armazenaro os clientes que estão escutando a estação
-#[derive(PartialEq, Clone, Debug)]
-pub struct Subscriber {
-    
-}
